pages folder moved under src
