+++ cc-ui/src/pages/Login.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
 
+//import app from './../cc-pk-service/src/app'
+
 import { 
     Heading,
     Container,
